classes: openrc: workaround file-rdeps
Workaround file-rdeps finding the shebang line in openrc initscripts and
complaining if openrc is not in RDEPENDS.  As openrc is being provided
as an alternative (and isn't supported directly upstream) that direct
RDEPEND on openrc isn't really desired.  Those wishing to use openrc can
add the package to their image directly.