classes: openrc: fix race in file-rdeps workaround
do_package_qa and do_package_write_* can both run at the same time.
This leads to a race in the restoration of the executable bit on openrc
init scripts.  Update to have a separate task that runs after
do_package_qa but before any of the do_package_write_* tasks.