openrc-scripts: remove
This recipe is going to become obsolete over the next series of commits
and it's easier to just remove it now rather then porting it over each
change just to remove it at the end once all functionality is
replicated.

The next series of commits will change this layer to behave more like
the update-rc.d/systemd classes in oe-core and use appends for recipes
where we provide openrc init.d scripts.

Signed-off-by: Justin Bronder <jsbronder@cold-front.org>