layer: refactor to behave like systemd bbclass
Rework the openrc recipe and bbclass to function similarly to the
systemd bbclass in openembedded-core.  Recipes can define
OPENRC_PACKAGES, each of which will have one or more OPENRC_SERVICES.
The class adds openrc-specific postinst and prerm hooks, updates
FILES:[package] as necessary, cleans out any legacy sysvinit scripts
and performs a few sanity checks.

As part of this work, go back to the openrc default of storing initd
scripts in /etc/init.d (instead of /etc/openrc.d).  Upstream never took
the patch that fixed overriding the default initdir in the Makefile and
has since changed to use meson.

Finally, update the openrc RDEPENDS so that all of the default
initscripts function out of the box.

Signed-off-by: Justin Bronder <jsbronder@cold-front.org>